Junit5 test 환경 세팅/테스트 완료

+++ android/app/src/main/java/com/downfall/caterplanner/common/model/Goal.java
@@ -12,11 +12,11 @@ public class Goal extends StatisticsDetailPlan{
 
     private Perform[] performs;
 
-    public Goal(long key, long headerId, int constructorKey, int constructorRelationType, String name, Type type, LocalDate startDate, LocalDate endDate, Integer hopeAchievement, String color, String cycle, int stat) {
+    public Goal(int key, Long headerId, int constructorKey, int constructorRelationType, String name, Type type, LocalDate startDate, LocalDate endDate, Integer hopeAchievement, String color, String cycle, int stat) {
         super(key, headerId, constructorKey, constructorRelationType, name, type, startDate, endDate, hopeAchievement, color, cycle, stat);
     }
 
-    public Goal(long key, long headerId, int constructorKey, int constructorRelationType, String name, Type type, LocalDate startDate, LocalDate endDate, Integer hopeAchievement, String color, String cycle, int stat, Perform[] performs) {
+    public Goal(int key, Long headerId, int constructorKey, int constructorRelationType, String name, Type type, LocalDate startDate, LocalDate endDate, Integer hopeAchievement, String color, String cycle, int stat, Perform[] performs) {
         super(key, headerId ,constructorKey, constructorRelationType, name, type, startDate, endDate, hopeAchievement, color, null, stat);
 
         this.performs = performs;
