+++ android/app/src/main/java/com/downfall/caterplanner/common/model/StatisticsDetailPlan.java
@@ -12,7 +12,7 @@ public abstract class StatisticsDetailPlan extends DetailPlan implements Briefin
     protected int currentBriefingCount;
     protected boolean isStatizable;
 
-    public StatisticsDetailPlan(long key, long headerId,  int constructorKey, int constructorRelationType, String name, Type type, LocalDate startDate, LocalDate endDate, Integer hopeAchievement, String color, String cycle, int stat) {
+    public StatisticsDetailPlan(int key, long headerId,  int constructorKey, int constructorRelationType, String name, Type type, LocalDate startDate, LocalDate endDate, Integer hopeAchievement, String color, String cycle, int stat) {
         super(key, headerId, constructorKey, constructorRelationType, name, type, startDate, endDate, hopeAchievement, color, cycle, stat);
     }
